PWN-5103 - remove last dot in input
