+++ ui-kit/src/main/kotlin/org/p2p/uikit/textwatcher/AmountFractionTextWatcher.kt
@@ -4,7 +4,6 @@ import android.text.Editable
 import android.text.TextWatcher
 import android.widget.EditText
 import org.p2p.uikit.R
-import org.p2p.uikit.utils.DecimalFormatter
 import org.p2p.uikit.utils.emptyString
 import org.p2p.uikit.utils.orZero
 import java.lang.ref.WeakReference
@@ -15,17 +14,15 @@ import kotlin.properties.Delegates
  * We prohibit entering the value after the dot if it's exceeded the max allowed length
  * For example: 10000.1234456789123123 -> 10 000.123456789
  * */
-
-private const val SYMBOL_ZERO = "0"
-private const val SYMBOL_DOT = "."
-private const val MAX_INT_LENGTH = 12
 private const val MAX_FRACTION_LENGTH = 9
 
 class AmountFractionTextWatcher(
     editText: EditText,
-    private val maxLengthAllowed: Int,
+    maxLengthAllowed: Int,
     private val onValueChanged: (String) -> Unit
 ) : TextWatcher {
+
+    private val amountFractionFormatter = AmountFractionFormatter(maxLengthAllowed)
 
     companion object {
         fun installOn(
@@ -63,17 +60,7 @@ class AmountFractionTextWatcher(
         // Move cursor one char left if user tries to remove space
         if (spaceWasRemoved) symbolsAfterCursor++
 
-        valueText = when {
-            value.isEmpty() -> value
-            // Remove whole string if dot before zero was removed
-            value == SYMBOL_ZERO && before == 1 -> emptyString()
-            value == "$SYMBOL_ZERO$SYMBOL_ZERO" && start == 1 -> SYMBOL_ZERO
-            value.startsWith(SYMBOL_DOT) -> "$SYMBOL_ZERO$value"
-            value.endsWith(SYMBOL_DOT) -> handleEndWithDotCase(value)
-            value.contains(SYMBOL_DOT) -> handleValueWithDotCase(value)
-            else -> handleGeneralCase(value)
-        }
-
+        valueText = amountFractionFormatter.formatAmountFraction(value, before, start)
         // Keep cursor in same place from the end
         cursorPosition = if (valueText.length < symbolsAfterCursor) 0 else valueText.length - symbolsAfterCursor
     }
@@ -89,34 +76,7 @@ class AmountFractionTextWatcher(
         }
     }
 
-    private fun handleEndWithDotCase(value: String): String = value.dropLast(1)
-        .dropSpaces()
-        .formatDecimal(maxLengthAllowed) + SYMBOL_DOT
-
-    private fun handleGeneralCase(value: String): String = value.dropSpaces()
-        .take(MAX_INT_LENGTH)
-        .formatDecimal(maxLengthAllowed)
-
-    private fun handleValueWithDotCase(value: String): String {
-        val dotPosition = value.indexOf(SYMBOL_DOT)
-        val intPart = value.substring(0, dotPosition)
-            .dropSpaces()
-            .take(MAX_INT_LENGTH)
-
-        // Remove extra dots and shorten fractional part to maxLengthAllowed symbols
-        val fractionalPart = value.substring(dotPosition + 1)
-            .dropSpaces()
-            .replace(SYMBOL_DOT, emptyString())
-            .take(maxLengthAllowed)
-
-        return intPart.formatDecimal(maxLengthAllowed) + SYMBOL_DOT + fractionalPart
-    }
-
     private fun String.dropSpaces(): String {
         return replace(" ", emptyString())
-    }
-
-    private fun String.formatDecimal(fractionLength: Int): String {
-        return DecimalFormatter.format(this.toBigDecimal(), fractionLength)
     }
 }
